pyproject.toml
[tool.poetry]
name = "voter"
version = "0.1.0"
description = ""
authors = ["Danzel88 <gecnepago@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
aiofiles = "^0.7.0"
SQLAlchemy = "^1.4.25"
psycopg2-binary = "^2.9.1"
email-validator = "^1.1.3"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
PyJWT = "^2.2.0"
asyncpg = "^0.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
